lab2_server/requirements.txt
python-dotenv==1.0.1
kafka-python==2.0.2
redis==5.0.7
faker==24.11.0
